refactor: extract dulicate entity ID resolution logic to shared helper
